requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
asyncio-mqtt==0.13.0
aiofiles==23.2.1

# Data validation and serialization
pydantic[email]==2.5.0

# Logging and monitoring
structlog==23.2.0

# CORS middleware (included in FastAPI but explicit version)
starlette==0.27.0

# Google Cloud and AI dependencies
google-generativeai==0.8.5
google-cloud-aiplatform==1.106.0
